+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # avromatic changelog
 
 ## v0.21.0
-- Add support for additional parameters to
-  `AvroTurf::ConfluentSchemaRegistry#compatible?`.
+- Remove monkey-patches for `AvroTurf::ConfluentSchemaRegistry` and
+  `FakeConfluentSchemaRegistryServer` and depend on `avro_schema_registry-client`
+  instead.
+- Rename the configuration option `use_cacheable_schema_registration` to
+  `use_schema_fingerprint_lookup`.
 
 ## v0.20.0
 - Support schema stores with a `#clear_schemas` method.